+++ src/components/content/observations/helpers/columns.tsx
@@ -1,12 +1,8 @@
-import { faChevronRight } from "@fortawesome/free-solid-svg-icons";
-import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
-import cx from "classnames";
-import { useState } from "react";
 import { BubbleIcon } from "./bubble";
+import { ObservationTrends } from "./trends";
 import { TableColumn } from "@/components/core/table/table-helpers";
 import { DiagnosticReportModel, ObservationModel } from "@/fhir/models";
 import { capitalize } from "@/utils/nodash";
-import { Telemetry } from "@/utils/telemetry";
 
 export const diagnosticReportColumns: TableColumn<DiagnosticReportModel>[] = [
   {
@@ -39,100 +35,40 @@ export const diagnosticReportColumns: TableColumn<DiagnosticReportModel>[] = [
   },
 ];
 
-export const ObservationsColumns = (): TableColumn<ObservationModel>[] => {
-  const [isTrendsShown, setIsTrendsShown] = useState(false);
-  return [
-    {
-      title: "Component",
-      render: (model) => (
-        <div className="ctw-flex ctw-flex-wrap ctw-text-base">{model.display}</div>
-      ),
-    },
-    {
-      title: "Result",
-      render: (model) => (
-        <div className="ctw-flex">
-          {model.interpretation
-            ? model.interpretation &&
-              model.value && (
-                <div className="ctw-text-sm ctw-font-normal">
-                  Result:{" "}
-                  <BubbleIcon
-                    interpretation={model.interpretation}
-                    result={model.value}
-                    className={model.acceptedInterpretations}
-                  />
-                </div>
-              )
-            : model.value && (
-                <div className="ctw-text-sm ctw-font-normal">
-                  Result:{" "}
-                  <BubbleIcon result={model.value} className={model.acceptedInterpretations} />
-                </div>
-              )}
-        </div>
-      ),
-    },
-    {
-      title: "Reference Range",
-      render: (model) => (
-        <div className="ctw-flex ctw-text-sm ctw-font-normal">
-          {model.referenceRange && (
-            <div className="ctw-font-normal">
-              Reference Range: {model.referenceRange} {model.unit}
-            </div>
-          )}
-        </div>
-      ),
-    },
-    {
-      title: "Trends",
-      render: (model) => (
+export const ObservationsColumns = (): TableColumn<ObservationModel>[] => [
+  {
+    title: "Component",
+    render: (model) => <div className="ctw-flex ctw-flex-wrap ctw-text-base">{model.display}</div>,
+  },
+  {
+    title: "Result",
+    render: (model) => (
+      <div className="ctw-flex">
         <div className="ctw-text-sm ctw-font-normal">
-          {model.trends.length > 1 && (
-            <div className="ctw-pt-2">
-              <button
-                aria-label="trends"
-                className="ctw-btn-clear"
-                type="button"
-                onClick={() => {
-                  Telemetry.trackInteraction(
-                    isTrendsShown ? "CollapsedTrendNotShown" : "ExpandedTrendShown"
-                  );
-                  setIsTrendsShown(!isTrendsShown);
-                }}
-              >
-                <FontAwesomeIcon
-                  icon={faChevronRight}
-                  className={cx("ctw-h-3 ctw-w-3 ctw-pr-1", {
-                    "ctw-rotate-90": isTrendsShown,
-                  })}
-                />
-                Trends
-              </button>
-              {isTrendsShown &&
-                model.trends.map((trend) => (
-                  <div key={trend.id} className="ctw-ml-4 ctw-py-px">
-                    <div
-                      className={cx("ctw-relative ctw-top-1 ctw-float-left ctw-w-24 ctw-text-sm", {
-                        "ctw-font-bold": trend.id === model.id,
-                      })}
-                    >
-                      {trend.effectiveStart}
-                    </div>
-                    <BubbleIcon
-                      result={trend.value}
-                      interpretation={trend.interpretation}
-                      className={cx(`${trend.acceptedInterpretations}`, {
-                        "!ctw-font-bold": trend.id === model.id,
-                      })}
-                    />
-                  </div>
-                ))}
-            </div>
-          )}
+          Result:{" "}
+          <BubbleIcon
+            interpretation={model.interpretation}
+            result={model.value}
+            className={model.acceptedInterpretations}
+          />
         </div>
-      ),
-    },
-  ];
-};
+      </div>
+    ),
+  },
+  {
+    title: "Reference Range",
+    render: (model) => (
+      <div className="ctw-flex ctw-text-sm ctw-font-normal">
+        {model.referenceRange && (
+          <div className="ctw-font-normal">
+            Reference Range: {model.referenceRange} {model.unit}
+          </div>
+        )}
+      </div>
+    ),
+  },
+  {
+    title: "Trends",
+    render: (model) => <ObservationTrends model={model} />,
+  },
+];