+++ src/components/core/simple-more-list.tsx
@@ -3,16 +3,23 @@ type SimpleMoreListProps = {
   items: string[];
   limit: number;
   total: number;
+  prefix?: string;
 };
 
-export const SimpleMoreList = ({ className, items, limit, total }: SimpleMoreListProps) => (
-  <div className={className}>
-    {items.slice(0, limit).map((item, index) => (
-      // eslint-disable-next-line react/no-array-index-key
-      <div key={item + index}>
-        <div>{item}</div>
-      </div>
-    ))}
-    {total > limit && <div className="ctw-font-medium">+ {total - limit} more</div>}
-  </div>
-);
+export const SimpleMoreList = ({ className, items, limit, total, prefix }: SimpleMoreListProps) => {
+  const displayItems = items;
+  if (prefix) {
+    displayItems[0] = `${prefix} ${displayItems[0]}`;
+  }
+  return (
+    <div className={className}>
+      {displayItems.slice(0, limit).map((item, index) => (
+        // eslint-disable-next-line react/no-array-index-key
+        <div key={item + index}>
+          <div>{item}</div>
+        </div>
+      ))}
+      {total > limit && <div className="ctw-font-medium">+ {total - limit} more</div>}
+    </div>
+  );
+};