+++ src/components/core/tab-group/tab-group.tsx
@@ -53,7 +53,7 @@ function TabGroupComponent({
   // https://github.com/tailwindlabs/headlessui/issues/2276#issuecomment-1456537475
   const [shown, setShown] = useState<Record<number, boolean>>({ 0: true });
 
-  // Calculate how many tabs can fit in the container.
+  // Calculate how many tabs can fit in the container,
   useEffect(() => {
     // Force everything into the more menu if we are in a small breakpoint
     // and we are not forcing horizontal tabs.
@@ -105,6 +105,7 @@ function TabGroupComponent({
   };
 
   const showTopRightContent = !!topRightContent;
+  const shownContent = content.slice(tabOverflowCutoff);
 
   return (
     <div
@@ -126,7 +127,7 @@ function TabGroupComponent({
           {content.map(({ key, display }, index) => (
             <Tab
               key={key}
-              data-zus-telemetry-click={`Tab[${key}]`}
+              data-zus-telemetry-click={`open_tab.${key}`}
               onClick={onClickBlur}
               className={({ selected }) =>
                 cx(
@@ -152,16 +153,19 @@ function TabGroupComponent({
               ref={moreMenuRef}
               selectedIndex={selectedTabIndex - tabOverflowCutoff}
               btnClassName={cx(
-                "ctw-more-tab ctw-capitalize ctw-flex-shrink-0 ctw-h-full",
+                "ctw-more-tab ctw-text-sm ctw-whitespace-nowrap ctw-capitalize ctw-flex-shrink-0 ctw-h-full",
                 "hover:after:ctw-bg-content-black",
                 {
                   "-ctw-ml-5": tabOverflowCutoff === 0,
                   "after:ctw-bg-content-black": selectedTabIndex - tabOverflowCutoff >= 0,
                 }
               )}
-              optionsClassName="ctw-tab-list ctw-capitalize"
+              optionsClassName={cx(
+                "ctw-tab-list ctw-capitalize",
+                !showTopRightContent && shownContent.length > 1 && "ctw-right-0" // Right-align dropdown if it is rightmost (and not leftmost)
+              )}
               onChange={(index) => handleOnChange(index + tabOverflowCutoff)}
-              items={content.slice(tabOverflowCutoff)}
+              items={shownContent}
             >
               {tabOverflowCutoff !== 0 ? <span>More</span> : undefined}
             </ListBox>