+++ src/index.ts
@@ -12,6 +12,8 @@ export * from "@/components/content/diagnostic-reports/patient-diagnostic-report
 export * from "@/components/content/diagnostic-reports/unread-diagnostic-reports-notification";
 export * from "@/components/content/document/patient-documents";
 export * from "@/components/content/document/unread-documents-notification";
+export * from "@/components/content/forms/actions/patients";
+export * from "@/components/content/encounters/patient-encounters";
 export * from "@/components/content/immunizations/patient-immunizations";
 export * from "@/components/content/immunizations/unread-immunizations-notification";
 export * from "@/components/content/medications/patient-medications";
@@ -27,6 +29,7 @@ export * from "@/components/content/patients/patients-search";
 export * from "@/components/content/patients/patients-table";
 export * from "@/components/content/timeline/patient-timeline";
 export * from "@/components/content/zus-aggregated-profile/zus-aggregated-profile";
+export * from "@/components/content/zus-aggregated-profile/zus-aggregated-profile-iframe";
 export * from "@/components/content/unread-records-notification";
 // Core components
 export * from "@/components/core/action-list/action-list";