+++ src/utils/telemetry.ts
@@ -48,8 +48,9 @@ const ignoredOrigins = [
 
 // Assume local development if origin is localhost or just an IP address
 const isLocalDevelopment = /https?:\/\/(localhost|\d+\.\d+\.\d+\.\d+)/i.test(origin);
-// Avoid initializing telemetry multiple times
+// Avoid initializing telemetry or event listeners multiple times
 let isInitialized = false;
+let listenersHaveBeenAdded = false;
 
 /**
  * Bootstrap APM Tracing
@@ -104,13 +105,9 @@ export class Telemetry {
     this.setEnv(environment);
     this.ehr = ehr;
 
-    if (this.telemetryIsAvailable) {
-      return;
-    }
-
     // Turning on Datadog Logging is conditional. However, the event handlers
     // that explicitly send internal /report/metrics to ctw are not optional.
-    if (allowDataDogLogging) {
+    if (!this.telemetryIsAvailable && allowDataDogLogging) {
       datadogLogs.init({
         ...(isLocalDevelopment ? devDatadogConfig : prodDatadogConfig),
         env: this.environment,
@@ -121,35 +118,39 @@ export class Telemetry {
       });
     }
 
-    // We are listening to click events propagating to the document body as that
-    // is the lowest level HTMLElement we actually know will both exist at this
-    // time and which won't be removed from the DOM by React causing a small
-    // memory leak.
-    // Additionally, because we aren't listening directly to events on elements
-    // that have `data-zus-telemetry-*` attributes, we don't have the luxury of
-    // knowing whether an event triggered from inside one of these telemetry
-    // elements, so we'll have to traverse the DOM tree ourselves. To minimize
-    // this work we'll put a `depth` level and adjust it over time if needed.
-    // For example, imagine the user clicked an element with CSS selector path
-    // of "button[data-zus-telemetry-click="submit"] > span > span". Because we
-    // are listening on body and not button, we would have to walk up 2 parent
-    // nodes of the DOM before knowing whether this event was relevant to us.
-    body?.addEventListener("click", (event) => {
-      const { target, isTrusted } = event;
-      if (!(isTrusted && target instanceof Element)) {
-        return;
-      }
-      const htmlElement = this.closestHTMLElement(target);
-      if (htmlElement instanceof HTMLElement) {
-        this.processHTMLEvent(htmlElement, "zusTelemetryClick");
-      }
-    });
-    body?.addEventListener("focusin", (event) => {
-      const { target, isTrusted } = event;
-      if (isTrusted && target instanceof HTMLElement) {
-        this.processHTMLEvent(target, "zusTelemetryFocus");
-      }
-    });
+    // We need to ensure that this will run in browser context, not just server.
+    if (!listenersHaveBeenAdded && body) {
+      // We are listening to click events propagating to the document body as that
+      // is the lowest level HTMLElement we actually know will both exist at this
+      // time and which won't be removed from the DOM by React causing a small
+      // memory leak.
+      // Additionally, because we aren't listening directly to events on elements
+      // that have `data-zus-telemetry-*` attributes, we don't have the luxury of
+      // knowing whether an event triggered from inside one of these telemetry
+      // elements, so we'll have to traverse the DOM tree ourselves. To minimize
+      // this work we'll put a `depth` level and adjust it over time if needed.
+      // For example, imagine the user clicked an element with CSS selector path
+      // of "button[data-zus-telemetry-click="submit"] > span > span". Because we
+      // are listening on body and not button, we would have to walk up 2 parent
+      // nodes of the DOM before knowing whether this event was relevant to us.
+      body.addEventListener("click", (event: MouseEvent) => {
+        const { target, isTrusted } = event;
+        if (!(isTrusted && target instanceof Element)) {
+          return;
+        }
+        const htmlElement = this.closestHTMLElement(target);
+        if (htmlElement instanceof HTMLElement) {
+          this.processHTMLEvent(htmlElement, "zusTelemetryClick");
+        }
+      });
+      body.addEventListener("focusin", (event) => {
+        const { target, isTrusted } = event;
+        if (isTrusted && target instanceof HTMLElement) {
+          this.processHTMLEvent(target, "zusTelemetryFocus");
+        }
+      });
+      listenersHaveBeenAdded = true;
+    }
 
     isInitialized = true;
   }
@@ -225,6 +226,16 @@ export class Telemetry {
       .replace(/[^a-z0-9_.]/gi, "");
   }
 
+  /*
+   * In dev/test environments we should skip sending any metrics
+   */
+  static skipMetrics() {
+    return (
+      !this.environment ||
+      (process.env.NODE_ENV !== "test" && /(localhost|127\.0\.0\.1)/.test(window.location.origin))
+    );
+  }
+
   /**
    * Report a metric to CTW
    */
@@ -234,7 +245,7 @@ export class Telemetry {
     value: number,
     additionalTags: string[] = []
   ) {
-    if (!this.isAllowedToSendMetrics) {
+    if (this.skipMetrics()) {
       return;
     }
 
@@ -378,7 +389,7 @@ export class Telemetry {
 
   static async reportActiveSession() {
     // Return if the session is already active or we don't need to report
-    if (Session.isActive() || !this.isAllowedToSendMetrics) {
+    if (Session.isActive() || this.skipMetrics()) {
       return;
     }